+++ jhotdraw7/src/main/java/org/jhotdraw/color/HSLHarmonicColorWheelImageProducer.java
@@ -1,12 +1,10 @@
 /**
  * @(#)HSLHarmonicColorWheelImageProducer.java
  *
- * Copyright (c) 2008 by the original authors of JHotDraw and all its
- * contributors. All rights reserved.
+ * Copyright (c) 2008 The authors and contributors of JHotDraw.
  *
  * You may not use, copy or modify this file, except in compliance with the
- * license agreement you entered into with the copyright holders. For details
- * see accompanying license terms.
+ * accompanying license terms.
  */
 package org.jhotdraw.color;
 
@@ -169,7 +167,7 @@ public class HSLHarmonicColorWheelImageProducer extends PolarColorWheelImageProd
         float radius = Math.min(w, h) / 2f;
 
         float[] hsb;
-        float sat = (float) r / radius;
+        float sat = r / radius;
         float hue = (float) (theta / Math.PI / 2d);
         if (hue < 0) {
             hue += 1f;
