+++ jhotdraw7/src/main/java/org/jhotdraw/color/PolarColorWheelImageProducer.java
@@ -1,12 +1,9 @@
 /*
  * @(#)PolarColorWheelImageProducer.java
  *
- * Copyright (c) 2008 by the original authors of JHotDraw and all its
- * contributors. All rights reserved.
- *
+ * Copyright (c) 2008 The authors and contributors of JHotDraw.
  * You may not use, copy or modify this file, except in compliance with the 
- * license agreement you entered into with the copyright holders. For details
- * see accompanying license terms.
+ * accompanying license terms.
  */
 package org.jhotdraw.color;
 
@@ -130,7 +127,7 @@ public class PolarColorWheelImageProducer extends AbstractColorWheelImageProduce
         float theta = (float) Math.atan2(y, -x);
 
         float angular = (float) (0.5 + (theta / Math.PI / 2d));
-        float radial=Math.min(1f, (float) r / getRadius());
+        float radial=Math.min(1f, r / getRadius());
 
         float[] hsb = new float[3];
         hsb[angularIndex] = angular//
