+++ jhotdraw7/src/main/java/org/jhotdraw/geom/Geom.java
@@ -1,12 +1,9 @@
 /*
  * @(#)Geom.java
  *
- * Copyright (c) 1996-2010 by the original authors of JHotDraw and all its
- * contributors. All rights reserved.
- *
+ * Copyright (c) 1996-2010 The authors and contributors of JHotDraw.
  * You may not use, copy or modify this file, except in compliance with the 
- * license agreement you entered into with the copyright holders. For details
- * see accompanying license terms.
+ * accompanying license terms.
  */
 package org.jhotdraw.geom;
 
@@ -94,8 +91,8 @@ public class Geom {
             return (abs(py - y1) <= tolerance);
         }
 
-        a = (double) (y1 - y2) / (double) (x1 - x2);
-        b = (double) y1 - a * (double) x1;
+        a = (y1 - y2) / (x1 - x2);
+        b = y1 - a * x1;
         x = (py - b) / a;
         y = a * px + b;
